+++ src/User/RegisterUserCommand.php
@@ -58,6 +58,9 @@ final class RegisterUserCommand
     /** @var string|null */
     private $comment;
 
+    /** @var string|null */
+    private $legalIdentifier;
+
     /** @return string|null */
     public function getPhone(): ?string
     {
@@ -338,6 +341,24 @@ final class RegisterUserCommand
     public function setComment(?string $comment): self
     {
         $this->comment = $comment;
+
+        return $this;
+    }
+
+    /** @return string|null */
+    public function getLegalIdentifier(): ?string
+    {
+        return $this->legalIdentifier;
+    }
+
+    /**
+     * @param string|null $legalIdentifier
+     *
+     * @return $this
+     */
+    public function setLegalIdentifier(?string $legalIdentifier): self
+    {
+        $this->legalIdentifier = $legalIdentifier;
 
         return $this;
     }